UI v1 is done except filter, I also applied some changes to fix some bugs.

+++ ResumableFunctions.Handler/Core/FirstWaitProcessor.cs
@@ -1,4 +1,5 @@
-﻿using System.Reflection;
+﻿using System.ComponentModel;
+using System.Reflection;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using ResumableFunctions.Handler.Core.Abstraction;
@@ -6,29 +7,32 @@ using ResumableFunctions.Handler.DataAccess;
 using ResumableFunctions.Handler.DataAccess.Abstraction;
 using ResumableFunctions.Handler.Helpers;
 using ResumableFunctions.Handler.InOuts;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace ResumableFunctions.Handler.Core;
 
 internal class FirstWaitProcessor : IFirstWaitProcessor
 {
     private readonly ILogger<FirstWaitProcessor> _logger;
-    private readonly FunctionDataContext _context;
+    private readonly IUnitOfWork _context;
     private readonly IMethodIdsRepo _methodIdentifierRepo;
     private readonly IWaitsRepo _waitsRepository;
+    private readonly IWaitTemplatesRepo _templatesRepo;
     private readonly IServiceProvider _serviceProvider;
     private readonly BackgroundJobExecutor _backgroundJobExecutor;
-    private readonly IResumableFunctionsSettings _settings;
     private readonly IBackgroundProcess _backgroundJobClient;
+    private readonly IServiceRepo _serviceRepo;
 
     public FirstWaitProcessor(
         ILogger<FirstWaitProcessor> logger,
-        FunctionDataContext context,
+        IUnitOfWork context,
         IServiceProvider serviceProvider,
         IMethodIdsRepo methodIdentifierRepo,
         IWaitsRepo waitsRepository,
         BackgroundJobExecutor backgroundJobExecutor,
-        IResumableFunctionsSettings settings,
-        IBackgroundProcess backgroundJobClient)
+        IBackgroundProcess backgroundJobClient,
+        IServiceRepo serviceRepo,
+        IWaitTemplatesRepo templatesRepo)
     {
         _logger = logger;
         _context = context;
@@ -36,8 +40,9 @@ internal class FirstWaitProcessor : IFirstWaitProcessor
         _methodIdentifierRepo = methodIdentifierRepo;
         _waitsRepository = waitsRepository;
         _backgroundJobExecutor = backgroundJobExecutor;
-        _settings = settings;
         _backgroundJobClient = backgroundJobClient;
+        _serviceRepo = serviceRepo;
+        _templatesRepo = templatesRepo;
     }
 
     public async Task<MethodWait> CloneFirstWait(MethodWait firstMatchedMethodWait)
@@ -45,7 +50,7 @@ internal class FirstWaitProcessor : IFirstWaitProcessor
         var rootId = int.Parse(firstMatchedMethodWait.Path.Split('/', StringSplitOptions.RemoveEmptyEntries)[0]);
         var resumableFunction =
             rootId != firstMatchedMethodWait.Id ?
-            (await _context.Waits.Include(x => x.RequestedByFunction).FirstAsync(x => x.Id == rootId)).RequestedByFunction.MethodInfo:
+            await _waitsRepository.GetRequestedByMethodInfo(rootId) :
             firstMatchedMethodWait.RequestedByFunction.MethodInfo;
 
         try
@@ -63,20 +68,19 @@ internal class FirstWaitProcessor : IFirstWaitProcessor
                         () => new LocalRegisteredMethods().TimeWait(
                         new TimeWaitInput
                         {
-                            TimeMatchId = firstMatchedMethodWait.MandatoryPart
+                            TimeMatchId = firstMatchedMethodWait.MandatoryPart,
+                            Description = $"`{timeWait.Name}` in function `{firstMatchedMethodWait.RequestedByFunction.RF_MethodUrn}:{firstMatchedMethodWait.FunctionState.Id}`"
                         }), timeWait.TimeToWait);
                     timeWait.TimeWaitMethod.MandatoryPart = firstMatchedMethodWait.MandatoryPart;
                     timeWait.IgnoreJobCreation = true;
                 }
 
             });
-            //firstWaitClone.FunctionState.AddLog(
-            //    $"[{resumableFunction.GetFullName()}] started and wait [{firstMatchedMethodWait.Name}] to match.", LogType.Info);
 
             firstWaitClone.FunctionState.Logs.AddRange(firstWaitClone.FunctionState.Logs);
             firstWaitClone.FunctionState.Status =
                 firstWaitClone.FunctionState.HasErrors() ?
-                FunctionStatus.Error :
+                FunctionStatus.InError :
                 FunctionStatus.InProgress;
             await _waitsRepository.SaveWait(firstWaitClone);//first wait clone
 
@@ -84,10 +88,7 @@ internal class FirstWaitProcessor : IFirstWaitProcessor
             currentMw.Status = WaitStatus.Waiting;
             currentMw.Input = firstMatchedMethodWait.Input;
             currentMw.Output = firstMatchedMethodWait.Output;
-            var waitTemplate = await _context
-                .WaitTemplates
-                .Select(WaitTemplate.BasicMatchSelector)
-                .FirstAsync(x => x.Id == firstMatchedMethodWait.TemplateId);
+            var waitTemplate = await _templatesRepo.GetWaitTemplateWithBasicMatch(firstMatchedMethodWait.TemplateId);
             currentMw.TemplateId = waitTemplate.Id;
             currentMw.Template = waitTemplate;
             currentMw.IsFirst = false;
@@ -99,56 +100,53 @@ internal class FirstWaitProcessor : IFirstWaitProcessor
         catch (Exception ex)
         {
             var error = $"Error when try to clone first wait for function [{resumableFunction.GetFullName()}]";
-            await LogErrorToService(ex, error);
+            await _serviceRepo.AddErrorLog(ex, error, ErrorCodes.FirstWait);
             throw new Exception(error, ex);
         }
     }
 
+    [DisplayName("Register First Wait for Function `{0}`")]
     public async Task RegisterFirstWait(int functionId)
     {
         MethodInfo resumableFunction = null;
-        var errorMsg = $"Error when try to register first wait for function [{functionId}]";
+        var functionName = "";
         await _backgroundJobExecutor.Execute(
             $"FirstWaitProcessor_RegisterFirstWait_{functionId}",
-            async () =>
+            (Func<Task>)(async () =>
             {
                 try
                 {
                     var resumableFunctionId = await _methodIdentifierRepo.GetResumableFunction(functionId);
                     resumableFunction = resumableFunctionId.MethodInfo;
-                    WriteMessage("START RESUMABLE FUNCTION AND REGISTER FIRST WAIT");
+                    functionName = resumableFunction.Name;
+                    _logger.LogInformation("START RESUMABLE FUNCTION AND REGISTER FIRST WAIT");
                     var firstWait = await GetFirstWait(resumableFunction, true);
+
                     if (firstWait != null)
-                        firstWait.FunctionState.AddLog(
+                    {
+                        await _serviceRepo.AddLog(
                             $"[{resumableFunction.GetFullName()}] started and wait [{firstWait.Name}] to match.");
-                    await _waitsRepository.SaveWait(firstWait);//first wait when register function
-                    WriteMessage($"Save first wait [{firstWait.Name}] for function [{resumableFunction.GetFullName()}].");
-                    await _context.SaveChangesAsync();
+
+                        await _waitsRepository.SaveWait(firstWait);
+                        _logger.LogInformation(
+                            $"Save first wait [{firstWait.Name}] for function [{resumableFunction.GetFullName()}].");
+                        await _context.SaveChangesAsync();
+                    }
                 }
                 catch (Exception ex)
                 {
                     if (resumableFunction != null)
-                        await LogErrorToService(ex, errorMsg);
+                        await _serviceRepo.AddErrorLog(ex, ErrorMsg(), ErrorCodes.FirstWait);
+
                     await _waitsRepository.RemoveFirstWaitIfExist(functionId);
                     throw;
                 }
 
-            },
-            errorMsg, true);
+            }),
+            ErrorMsg(), true);
+        string ErrorMsg() => $"Error when try to register first wait for function [{functionName}:{functionId}]";
     }
 
-    private async Task LogErrorToService(Exception ex, string errorMsg)
-    {
-        _logger.LogError(ex, errorMsg);
-        _context.Logs.Add(new LogRecord
-        {
-            EntityId = _settings.CurrentServiceId,
-            EntityType = nameof(ServiceData),
-            Message = errorMsg + ex,
-            Type = LogType.Error
-        });
-        await _context.SaveChangesAsync();
-    }
 
     public async Task<Wait> GetFirstWait(MethodInfo resumableFunction, bool removeIfExist)
     {
@@ -158,7 +156,8 @@ internal class FirstWaitProcessor : IFirstWaitProcessor
         {
 
             var errorMsg = $"Can't initiate a new instance of [{resumableFunction.DeclaringType.FullName}]";
-            await LogErrorToService(null, errorMsg);
+            await _serviceRepo.AddErrorLog(null, errorMsg, ErrorCodes.FirstWait);
+
             throw new NullReferenceException(errorMsg);
         }
 
@@ -169,16 +168,24 @@ internal class FirstWaitProcessor : IFirstWaitProcessor
         {
             var message = $"Resumable function ({resumableFunction.GetFullName()}) not exist in code.";
             _logger.LogWarning(message);
-            await LogErrorToService(null, message);
+            await _serviceRepo.AddErrorLog(null, message, ErrorCodes.FirstWait);
+
             throw new NullReferenceException(message);
         }
 
         await functionRunner.MoveNextAsync();
         var firstWait = functionRunner.Current;
+
+        if (firstWait == null)
+        {
+            await _serviceRepo.AddErrorLog(null, $"Can't get first wait in function `{resumableFunction.GetFullName()}`.", ErrorCodes.FirstWait);
+            return null;
+        }
+
         var methodId = await _methodIdentifierRepo.GetResumableFunction(new MethodData(resumableFunction));
         if (removeIfExist)
         {
-            WriteMessage("First wait already exist it will be deleted and recreated since it may be changed.");
+            _logger.LogInformation("First wait already exist it will be deleted and recreated since it may be changed.");
             await _waitsRepository.RemoveFirstWaitIfExist(methodId.Id);
         }
         var functionState = new ResumableFunctionState
@@ -197,38 +204,35 @@ internal class FirstWaitProcessor : IFirstWaitProcessor
         return firstWait;
     }
 
-    private void WriteMessage(string message)
-    {
-        _logger.LogInformation(message);
-    }
 
-    public async Task DeactivateFirstWait(int functionId)
-    {
-        await _backgroundJobExecutor.Execute(
-            $"DeactivateFirstWait_{functionId}",
-            async () =>
-            {
-                var firstWaits = await _context
-                        .Waits
-                        .Include(x => x.FunctionState)
-                        .Where(wait =>
-                                wait.RequestedByFunctionId == functionId &&
-                                wait.IsFirst &&
-                                wait.Status == WaitStatus.Waiting)
-                        .ToListAsync();
+    //public async Task DeactivateFirstWait(int functionId)
+    //{
+    //    await _backgroundJobExecutor.Execute(
+    //        $"DeactivateFirstWait_{functionId}",
+    //        async () =>
+    //        {
+    //            var firstWaits = await _context
+    //                    .Waits
+    //                    .Include(x => x.FunctionState)
+    //                    .Where(
+    //                        wait =>
+    //                        wait.RequestedByFunctionId == functionId &&
+    //                        wait.IsFirst &&
+    //                        wait.Status == WaitStatus.Waiting)
+    //                    .ToListAsync();
 
-                foreach (var firstWait in firstWaits)
-                {
-                    if (firstWait != default)
-                    {
-                        firstWait.IsFirst = false;
-                        firstWait.Cancel();
-                        _context.Waits.Remove(firstWait);
-                        _context.FunctionStates.Remove(firstWait.FunctionState);
-                    }
-                    await _context.SaveChangesAsync();
-                }
-            },
-            $"Error when try to deactivate first wait for function [{functionId}]", true);
-    }
+    //            foreach (var firstWait in firstWaits)
+    //            {
+    //                if (firstWait != default)
+    //                {
+    //                    firstWait.IsFirst = false;
+    //                    firstWait.Cancel();
+    //                    _context.Waits.Remove(firstWait);
+    //                    _context.FunctionStates.Remove(firstWait.FunctionState);
+    //                }
+    //                await _context.SaveChangesAsync();
+    //            }
+    //        },
+    //        $"Error when try to deactivate first wait for function `{functionId}`.", true);
+    //}
 }
